requirements.txt
anyio==4.9.0
asyncio==3.4.3
certifi==2025.4.26
charset-normalizer==3.4.2
Deprecated==1.2.18
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.6.1
microsoft-kiota-abstractions==1.9.3
microsoft-kiota-http==1.9.3
microsoft-kiota-serialization-form==1.9.3
microsoft-kiota-serialization-json==1.9.3
microsoft-kiota-serialization-multipart==1.9.3
microsoft-kiota-serialization-text==1.9.3
opentelemetry-api==1.33.0
opentelemetry-sdk==1.33.0
opentelemetry-semantic-conventions==0.54b0
requests==2.32.3
sniffio==1.3.1
std-uritemplate==2.0.5
typing_extensions==4.13.2
urllib3==2.4.0
wrapt==1.17.2
zipp==3.21.0
